Update noahClass.java
ok so this is cool i can either directly comit it to your branch right now or oi can make it into a new brnach and send you a pull request to pull it inot your thing. im going direct so you will see the changes and not get a request. ha.